+++ migrations/000001_create_users_table.down.sql
@@ -1,2 +1,2 @@
-DROP TABLE IF EXISTS users;
 DROP FUNCTION IF EXISTS update_updated_at;
+DROP TABLE IF EXISTS users;